# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger",
    "description": "Testen von verschiedenen GET/POST Methoden",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/find/users": {
      "parameters": [
        {
          "name": "_id",
          "in": "query",
          "required": false,
          "description": "ID des Benutzers"
        },
        {
          "name": "name",
          "in": "query",
          "required": false,
          "description": "Name des Benutzers"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User",
        "responses": {
          "200": {
            "description": "Benutzer wurde(n) gefunden",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/insert/users": {
      "parameters": [
        {
          "name": "name",
          "in": "query",
          "required": true,
          "description": "Name des Benutzers"
        }
      ],
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Insert User",
        "responses": {
          "200": {
            "description": "Benutzer wurde hinzugefügt",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/update/users/{_id}": {
      "parameters": [
        {
          "name": "_id",
          "in": "path",
          "required": true,
          "description": "ID des Benutzers"
        },
        {
          "name": "name",
          "in": "query",
          "required": false,
          "description": "Neuer Name des Benutzers"
        },
        {
          "name": "token",
          "in": "query",
          "required": false,
          "description": "Neuer Token des Benutzers"
        }
      ],
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Update User",
        "responses": {
          "200": {
            "description": "Benutzerinformation wurden aktualisiert",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/delete/users": {
      "parameters": [
        {
          "name": "_id",
          "in": "query",
          "required": true,
          "description": "ID des Benutzers"
        },
        {
          "name": "name",
          "in": "query",
          "required": false,
          "description": "Name des Benutzers"
        }
      ],
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Delete User",
        "responses": {
          "200": {
            "description": "Benutzer wurde gelöscht",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/find/scan/": {
      "parameters": [
        {
          "name": "identifier",
          "in": "query",
          "required": true,
          "description": "Raum-Identifier"
        }
      ],
      "get": {
        "tags": [
          "Scans"
        ],
        "summary": "Get ScanResult",
        "responses": {
          "200": {
            "description": "Scan war erfolgreich",
            "schema": {
              "$ref": "#/definitions/Scan"
            }
          }
        }
      }
    },
...
    "/post/answer": {
      "parameters": [
        {
          "name": "gameId",
          "in": "query",
          "required": true,
          "description": "ID des Spiels"
        },
        {
          "name": "answer",
          "in": "query",
          "required": true,
          "description": "Antwort zu dem Spiel"
        }
      ],
      "post": {
        "tags": [
          "GameAnswer"
        ],
        "summary": "Post Answer for an Quiz",
        "responses": {
          "200": {
            "description": "Die Antwort ist richtig"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": [
        "_id"
      ],
      "properties": {
        "_id": {
          "type": "integer",
          "uniqueItems": true
        },
        "name": {
          "type": "string",
          "uniqueItems": true
        },
        "token": {
          "type": "integer"
        },
        "visits": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Visit"
          }
        }
      }
    },
    "Location": {
      "required": [
        "_id"
      ],
      "properties": {
        "_id": {
          "type": "integer",
          "uniqueItems": true
        },
        "roomnumber": {
          "type": "string",
          "uniqueItems": true
        },
        "identifier": {
          "type": "string",
          "uniqueItems": true
        }
      }
    },
    "Game": {
      "required": [
        "answers",
        "question",
        "type",
        "points"
      ],
      "properties": {
        "question": {
          "type": "string"
        },
        "answers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Answer"
          }
        },
        "type": {
          "type": "string",
          "enum": [
            "single_choice",
            "multiple_choice"
          ]
        },
        "points": {
          "type": "integer"
        }
      }
    },
    "Answer": {
      "required": [
        "answer",
        "isCorrect"
      ],
      "properties": {
        "answer": {
          "type": "string"
        },
        "isCorrect": {
          "type": "boolean"
        }
      }
    },
...
    "Visit": {
      "required": [],
      "properties": {
        "location": {
          "$ref": "#/definitions/Location"
        },
        "games": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Game"
          }
        },
        "isSuccessful": {
          "type": "boolean"
        },
        "roomstate": {
          "type": "int"
        }
      }
    },
    "Scan": {
      "required": [],
      "properties": {
        "room": {
          "$ref": "#/definitions/Location"
        },
        "games": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Game"
          }
        }
      }
    },
...
  }
}